move long press announcement to useGrid and test
