+++ packages/@react-aria/grid/intl/en-US.json
@@ -3,5 +3,6 @@
   "select": "Select",
   "selectedCount": "{count, plural, =0 {No items selected} one {# item selected} other {# items selected}}.",
   "selectedAll": "All items selected.",
-  "selectedItem": "{item} selected."
+  "selectedItem": "{item} selected.",
+  "longPressToSelect": "Long press to enter selection mode."
 }
